+++ wolips/plugins/ng.componenteditor/java/org/objectstyle/wolips/apieditor/ApieditorPlugin.java
@@ -110,7 +110,7 @@ public class ApieditorPlugin extends AbstractBaseUIActivator {
 	 * @return the image descriptor
 	 */
 	public static ImageDescriptor getImageDescriptor(String path) {
-		return AbstractUIPlugin.imageDescriptorFromPlugin("org.objectstyle.wolips.apieditor", path);
+		return AbstractUIPlugin.imageDescriptorFromPlugin("ng.componenteditor", path);
 	}
 
 	public Image getImage(String key) {