+++ wolips/plugins/ng.componenteditor/java/org/objectstyle/wolips/apieditor/editor/ValidationPage.java
@@ -57,7 +57,7 @@ package org.objectstyle.wolips.apieditor.editor;
 
 public class ValidationPage extends ApiFormPage {
 
-	public static String PAGE_ID = "org.objectstyle.wolips.wodclipse.api.ValidationPage";
+	public static String PAGE_ID = "ng.componenteditor.api.ValidationPage";
 
 	public ValidationPage(ApiEditor apiEditor, String title) {
 		super(apiEditor, PAGE_ID, title);