+++ wolips/plugins/ng.componenteditor/java/org/objectstyle/wolips/templateeditor/TemplateEditor.java
@@ -51,7 +51,7 @@ import tk.eclipse.plugin.htmleditor.editors.HTMLEditorPart;
 import tk.eclipse.plugin.htmleditor.editors.HTMLSourceEditor;
 
 public class TemplateEditor extends HTMLEditor implements IEmbeddedEditor, IHtmlDocumentProvider, IWebobjectTagListener, ITextEditor, IReusableEditor, ITextEditorExtension, ITextEditorExtension2, ITextEditorExtension3, ITextEditorExtension4, INavigationLocationProvider, ISaveablesSource, IPersistableEditor, ITextWOEditor {
-  public static final String BINDING_HOVER_ANNOTATION = "org.objectstyle.wolips.tkhtmleditor.bindingHover";
+  public static final String BINDING_HOVER_ANNOTATION = "ng.componenteditor.bindingHover";
 
   private TemplateConfiguration _configuration;
   private ComponentEditorInteraction _editorInteraction;