+++ wolips/plugins/ng.componenteditor/java/org/objectstyle/wolips/wodclipse/WodclipsePlugin.java
@@ -104,7 +104,7 @@ public class WodclipsePlugin extends AbstractBaseUIActivator {
 	 * @return the image descriptor
 	 */
 	public static ImageDescriptor getImageDescriptor(String path) {
-		return AbstractUIPlugin.imageDescriptorFromPlugin("org.objectstyle.wolips.wodclipse", path);
+		return AbstractUIPlugin.imageDescriptorFromPlugin("ng.componenteditor", path);
 	}
 
 	public Image getImage(String key) {