+++ wolips/plugins/ng.componenteditor/java/org/objectstyle/wolips/wodclipse/core/document/WodFileDocumentProvider.java
@@ -59,7 +59,7 @@ import org.objectstyle.wolips.wodclipse.core.parser.IWodFilePartitions;
  * @author mike
  */
 public class WodFileDocumentProvider extends TextFileDocumentProvider {
-	public static final String WOD_FILE_CONTENT_TYPE_STRING = "org.objectstyle.wolips.wodclipse.wod";
+	public static final String WOD_FILE_CONTENT_TYPE_STRING = "ng.componenteditor.wod";
 
 	public static final IContentType WOD_FILE_CONTENT_TYPE = Platform.getContentTypeManager().getContentType(WodFileDocumentProvider.WOD_FILE_CONTENT_TYPE_STRING); //$NON-NLS-1$